% Auto-generated: do not edit by hand
\name{''GridItem}

\alias{''GridItem}

\title{GridItem component}

\description{
An individual item within a ReactGridLayout.
}

\usage{
''GridItem(children=NULL, cols=NULL, containerWidth=NULL, rowHeight=NULL, margin=NULL,
maxRows=NULL, containerPadding=NULL, x=NULL, y=NULL, w=NULL, h=NULL, minW=NULL,
maxW=NULL, minH=NULL, maxH=NULL, i=NULL, onDragStop=NULL, onDragStart=NULL,
onDrag=NULL, onResizeStop=NULL, onResizeStart=NULL, onResize=NULL,
isDraggable=NULL, isResizable=NULL, static=NULL, useCSSTransforms=NULL,
transformScale=NULL, className=NULL, handle=NULL, cancel=NULL,
droppingPosition=NULL, usePercentages=NULL, style=NULL)
}

\arguments{

}
