% Auto-generated: do not edit by hand
\name{''ReactGridLayout}

\alias{''ReactGridLayout}

\title{ReactGridLayout component}

\description{
A reactive, fluid grid layout with draggable, resizable components.
}

\usage{
''ReactGridLayout(autoSize=NULL, cols=NULL, className=NULL, style=NULL, draggableHandle=NULL,
draggableCancel=NULL, containerPadding=NULL, rowHeight=NULL, maxRows=NULL,
layout=NULL, margin=NULL, isDraggable=NULL, isResizable=NULL, isDroppable=NULL,
useCSSTransforms=NULL, transformScale=NULL, verticalCompact=NULL,
compactType=NULL, preventCollision=NULL, droppingItem=NULL, onLayoutChange=NULL,
onDragStart=NULL, onDrag=NULL, onDragStop=NULL, onResizeStart=NULL,
onResize=NULL, onResizeStop=NULL, onDrop=NULL)
}

\arguments{

}
