% Auto-generated: do not edit by hand
\name{myCalendar}

\alias{myCalendar}

\title{MyCalendar component}

\description{
MyCalendar allows the user to select a date
}

\usage{
myCalendar(id=NULL, date=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{date}{Character. The date as a string}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

