% Auto-generated: do not edit by hand
\name{myEditor}

\alias{myEditor}

\title{MyEditor component}

\description{
MyEditor is a quill editor.
}

\usage{
myEditor(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The edited content}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

