% Auto-generated: do not edit by hand
\name{''DragGrid}

\alias{''DragGrid}

\title{DragGrid component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DragGrid(children=NULL, id=NULL, label=NULL, divstyle=NULL, layout=NULL,
compacttype=NULL, rowheight=NULL, width=NULL, numcolumns=NULL, maxrows=NULL,
margin=NULL)
}

\arguments{

}
